Extract macro contents into new files
